requirements.txt
# Core FastAPI framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0

# Data validation
pydantic>=2.5.2
pydantic-settings>=2.1.0

# Environment variables
python-dotenv>=1.0.0

# File upload handling
python-multipart>=0.0.6

# Supabase Client (replaces SQLAlchemy - REST API only!)
supabase>=2.0.0
psycopg2-binary>=2.9.9  # PostgreSQL driver (for direct connections if needed)

# ML dependencies - Sentence Transformers for semantic similarity
sentence-transformers>=2.2.2
torch>=2.2.0

# Transformers for additional NLP tasks
transformers>=4.35.2

# HTTP requests for LLM API calls
requests>=2.31.0

# Grammar checking
language-tool-python>=2.7.1

# Resume parsing
PyPDF2>=3.0.1
python-docx>=1.1.0

# Audio processing
pydub>=0.25.1

# ===========================================
# Voice Analysis (librosa-based)
# ===========================================

# Librosa for advanced audio analysis (pitch, energy, rhythm)
librosa>=0.10.1

# Audio file I/O for librosa
soundfile>=0.12.1

# ===========================================
# Transcription (Speech-to-Text)
# ===========================================

# RECOMMENDED: Faster-Whisper for high-quality local transcription (no API key!)
# Uses CTranslate2 for 4x faster inference than OpenAI Whisper
faster-whisper>=0.10.0

# AssemblyAI for cloud transcription (has free tier)
assemblyai>=0.17.0

# OpenAI API (for transcription and LLM)
openai>=1.3.7

# Google Generative AI (for Gemini LLM - PRIMARY)
google-generativeai>=0.8.6

# ===========================================
# Database & Cloud (Supabase)
# ===========================================

# Supabase Python client (REST API - no database_url needed!)
supabase>=2.0.0

# JWT Authentication (for Supabase token validation)
python-jose[cryptography]>=3.3.0
PyJWT>=2.8.0

# Testing
pytest>=7.4.3
httpx>=0.25.2

# Additional utilities
numpy>=1.26.0
scipy>=1.11.0

# ML utilities
scikit-learn>=1.3.0
joblib>=1.3.0

# ===========================================
# PDF Report Generation
# ===========================================

# xhtml2pdf for HTML to PDF conversion (lightweight, pure Python)
xhtml2pdf>=0.2.11

# WeasyPrint for higher quality PDF (optional - requires system deps)
# weasyprint>=60.0  # Uncomment if system dependencies available

# Jinja2 for HTML templating
jinja2>=3.1.2

